Factor out export tests
